refactor project structur
